backend/requirements.txt
Django>=5.0
django-cors-headers
djangorestframework
psycopg[binary]
python-dotenv
drf-spectacular
channels[daphne]
celery
redis
channels_redis
